LastCheck

+++ admin/src/Page/DLICT.js
@@ -12,19 +12,23 @@ import jwtDecode from "jwt-decode";
 const { REACT_APP_PATH2 } = process.env;
 
 const DLICTPage = () => {
-    /////////////////////////////////////////////////////
+    ////////////////////////////////////
     const jwt = localStorage.getItem("mini-session");
     const navigate = useNavigate();
     if (!jwt) {
         navigate("/Login");
     }
-    const { exp } = jwtDecode(jwt);
-    const expirationTime = exp * 1000 - 60000;
-    if (Date.now() >= expirationTime) {
-        localStorage.clear();
-        navigate("/Login");
+
+    if (jwt) {
+        const { exp } = jwtDecode(jwt);
+        const expirationTime = exp * 1000 - 60000;
+        if (Date.now() >= expirationTime) {
+            localStorage.clear();
+            navigate("/Login");
+        }
     }
-    ////////////////////////////////////////////////////////////////
+    ////////////////////////////////////////////////////////////
+
     const type = "dlict";
     const [Data, setData] = useState();
     useEffect(() => {
