+++ admin/src/Page/Mission.js
@@ -5,26 +5,29 @@ import { Container } from "react-bootstrap";
 import { Spinner } from "react-bootstrap";
 import { ToastContainer } from "react-toastify";
 import ListGroup from "react-bootstrap/ListGroup";
-import SubMission from "../Page/SubPage/subMission"
+import SubMission from "../Page/SubPage/subMission";
 import { useNavigate } from "react-router-dom";
 import jwtDecode from "jwt-decode";
-const {REACT_APP_PATH2} = process.env;
+const { REACT_APP_PATH2 } = process.env;
 
 const Mission = () => {
-  
-/////////////////////////////////////////////////////
-const jwt = localStorage.getItem("mini-session");
-const navigate = useNavigate();
-if (!jwt) {
-    navigate("/Login");
-}
-const { exp } = jwtDecode(jwt);
-const expirationTime = exp * 1000 - 60000;
-if (Date.now() >= expirationTime) {
-    localStorage.clear();
-    navigate("/Login");
-}
-////////////////////////////////////////////////////////////////
+    ////////////////////////////////////
+    const jwt = localStorage.getItem("mini-session");
+    const navigate = useNavigate();
+    if (!jwt) {
+        navigate("/Login");
+    }
+
+    if (jwt) {
+        const { exp } = jwtDecode(jwt);
+        const expirationTime = exp * 1000 - 60000;
+        if (Date.now() >= expirationTime) {
+            localStorage.clear();
+            navigate("/Login");
+        }
+    }
+    ////////////////////////////////////////////////////////////
+
     const [Data, setData] = useState();
     useEffect(() => {
         function get() {
@@ -48,7 +51,7 @@ if (Date.now() >= expirationTime) {
                                     <ListGroup variant="flush">
                                         {Data ? (
                                             Data.map((data) => {
-                                                return  <SubMission key={data.id} data={data}/>
+                                                return <SubMission key={data.id} data={data} />;
                                             })
                                         ) : (
                                             <Spinner
