+++ admin/src/Page/Person.js
@@ -10,19 +10,23 @@ import { useNavigate } from "react-router-dom";
 import jwtDecode from "jwt-decode";
 const { REACT_APP_PATH2 } = process.env;
 const Person = () => {
-    /////////////////////////////////////////////////////
+    ////////////////////////////////////
     const jwt = localStorage.getItem("mini-session");
     const navigate = useNavigate();
     if (!jwt) {
         navigate("/Login");
     }
-    const { exp } = jwtDecode(jwt);
-    const expirationTime = exp * 1000 - 60000;
-    if (Date.now() >= expirationTime) {
-        localStorage.clear();
-        navigate("/Login");
+
+    if (jwt) {
+        const { exp } = jwtDecode(jwt);
+        const expirationTime = exp * 1000 - 60000;
+        if (Date.now() >= expirationTime) {
+            localStorage.clear();
+            navigate("/Login");
+        }
     }
-    ////////////////////////////////////////////////////////////////
+    ////////////////////////////////////////////////////////////
+
     const [Data, setData] = useState("");
     useEffect(() => {
         function get() {
